Working on creating and uploading unserialized bag. Post-upload deletion is not working.

+++ workers/job_runner.js
@@ -5,6 +5,7 @@ const { Context } = require('../core/context');
 const fs = require('fs')
 const { Job } = require('../core/job');
 const { OperationResult } = require('../core/operation_result');
+const path = require('path')
 const { Uploader } = require('./uploader');
 const { Util } = require('../core/util');
 const { ValidationOperation } = require('../core/validation_operation');
@@ -96,13 +97,17 @@ class JobRunner {
             this.assignUploadSources();
             let uploader = new Uploader(this.job);
             try {
+                console.log("STARTING")
                 await uploader.run();
+                console.log("DONE")
             } catch (ex) {
-                // console.error(ex);
+                console.log("ERROR")
+                console.error(ex);
                 // Note that the error will already be recorded in
                 // uploadOp.results[i].errors, and will be handled
                 // above like any other worker error.
             }
+            console.log("NEXT")
             var lastResult = null
             for (let op of this.job.uploadOps) {
                 for (let result of op.results) {
@@ -112,7 +117,11 @@ class JobRunner {
                     lastResult = result
                 }
             }
-            this.deleteBagAfterUpload(this.job, lastResult)
+            console.log(lastResult)
+            console.log(returnCode)
+            if (returnCode == Constants.EXIT_SUCCESS && this.job.deleteBagAfterUpload) {
+                this.deleteBagAfterUpload(this.job, lastResult)
+            }
             this.job.save();
             return returnCode;
         }
@@ -133,8 +142,20 @@ class JobRunner {
         let packOp = this.job.packageOp;
         for (let uploadOp of this.job.uploadOps) {
             if (packOp && packOp.outputPath && uploadOp.sourceFiles.length == 0) {
-                uploadOp.sourceFiles.push(packOp.outputPath);
+                if (Util.isDirectory(packOp.outputPath)) {
+                    // Add the directory and all its contents to the upload source files
+                    uploadOp.sourceFiles = Util.walkSync(packOp.outputPath).map((f) => f.absPath)
+                    for (let i=0; i < uploadOp.sourceFiles.length; i++) {
+                        let absPath = uploadOp.sourceFiles[i]
+                        uploadOp.sourceKeys[i] = absPath.replace(path.dirname(this.job.packageOp.outputPath) + path.sep, "")
+                    }
+                } else {
+                    // This is an individual file
+                    uploadOp.sourceFiles.push(packOp.outputPath);
+                }
             }
+            // console.log(uploadOp.sourceFiles)
+            // console.log(uploadOp.sourceKeys)
         }
     }
 
@@ -148,18 +169,26 @@ class JobRunner {
      * @returns {void}
      */
     deleteBagAfterUpload(job, result) {
+        console.log("1")
         if (!job.deleteBagAfterUpload) {
+            console.log("2")
             return
         }
+        console.log("3")
         let bag = job.packageOp.outputPath
+        console.log("4")
         try {
+            console.log("5")
             if (Util.isDirectory(bag)) {
-               Util.deleteRecursive(bag)
+                console.log("6")
+                Util.deleteRecursive(bag)
             } else {
-               fs.unlinkSync(bag)
+                console.log("7")
+                fs.unlinkSync(bag)
             }
             job.bagWasDeletedAfterUpload = true
         } catch(ex) {
+            console.error(ex)
             job.bagWasDeletedAfterUpload = false
             if (result) {
                 result.errors.push(`Upload completed but cound not delete bag ${bag}. Error: ${ex}`)
